Split JSX -> HAST into smaller pieces
